bug fix: tenser giving garbage values
tenser producing garbage values was not a bug related to indexing, but instead the difference between using a constructor initializer or not when allocating the data array

to address this, there is no longer a public constructor for tensers, but instead you have to use a static method to call the private constructor. this forces the user to use a certain contract when creating tensers.
you can either call tenser::garbage or tenser::defaults to pick if you want to initialize each item of type T in the to garbage or call the default constructors
specifying garbage vs default is important since T might be a class without a default constructor, so you can assign the items later