fixed NN passing the whole input into each neuron in the input layer + replacing double pointer methods with vector methods
the feed method in the NN class was passing the whole input array pointer into each neuron input layer. fixed this by making the current layer start at the first hidden layer rather than the input layer. we can delete the input layer and replace it with a variable later
also will replace the methods that accept pointers to arrays into vectors to insure bounds checking. the feed, train methods and output methods have been updated to account for this.